chore(asr/cohere): drop dead and buggy parallel pipelines
Tier A — fully dead code (zero callers anywhere in the tree):
- CohereStatelessManager.swift
- CohereModelInference.swift
- CohereDecoderState.swift (only referenced by CohereModelInference)

Tier B — "original buggy" pipeline that CohereFixedPipeline was written
to replace (per its own header comment), kept in parallel until now:
- CohereAsrManager.swift
- CohereAsrModels.swift (only consumer was CohereAsrManager + its CLI)
- CohereMelSpectrogram.swift (CohereFixedPipeline has its own internal
  CohereFixedMelSpectrogram; only the buggy manager + dead stateless
  manager + tests used this one)
- CohereTranscribeCommand.swift / CohereBenchmark.swift (only entries
  to the buggy manager)
- CohereMelSpectrogramTests.swift / CohereTokenConversionTests.swift

Lifted CohereAsrError into CohereFixedPipeline.swift (the lone surviving
consumer). Updated ModelNames.CohereTranscribe.requiredModels to point
at the cache-external compiled artifacts that actually ship; removed
the dangling decoderFile alias. Trimmed cohere-transcribe and
cohere-benchmark from the CLI dispatch and help text — cohere-mixed
and cohere-mixed-benchmark (which run the canonical CohereFixedPipeline,
the source of the published FLEURS numbers) remain.

Net: -2,584 deleted, ~28 added. swift build clean.
